# Parallelize Grunt build