--- a/package.json
+++ b/package.json
@@ -63,6 +63,7 @@
     "grunt-cli": "^1.4.3",
     "grunt-contrib-clean": "^2.0.1",
     "grunt-contrib-copy": "^1.0.0",
+    "grunt-parallel": "^0.5.1",
     "grunt-run": "^0.8.1",
     "grunt-timer": "^0.6.0",
     "grunt-ts": "^6.0.0-beta.22",