--- a/src/common/tools/gen_listings.ts
+++ b/src/common/tools/gen_listings.ts
@@ -9,7 +9,7 @@ function usage(rc: number): void {
 
 For each suite in SUITE_DIRS, generate listings and write each listing.js
 into OUT_DIR/{suite}/listing.js. Example:
-  tools/gen_listings out/ src/unittests/ src/webgpu/
+  tools/gen_listings gen/ src/unittests/ src/webgpu/
 
 Options:
   --help          Print this message and exit.
@@ -52,12 +52,5 @@ for (const suiteDir of argv.slice(3)) {
 export const listing = ${JSON.stringify(listing, undefined, 2)};
 `
     );
-
-    // If there was a sourcemap for the file we just replaced, delete it.
-    try {
-      fs.unlinkSync(outFile + '.map');
-    } catch (ex) {
-      // ignore if file didn't exist
-    }
   });
 }