… tools/gen_version